% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expandEpoch.R
\name{expandEpoch}
\alias{expandEpoch}
\title{Converts scrapePerson data w/ raw = TRUE into 30 or 60 epoch data}
\usage{
expandEpoch(data, epoch)
}
\arguments{
\item{data}{data frame from within scrapePerson with raw = TRUE}

\item{epoch}{currently only 30 is supported}
}
\value{
data frame converted to 30 or 60 second epoch
}
\description{
Converts scrapePerson data w/ raw = TRUE into 30 or 60 epoch data
}
