% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/importHelpers.R
\name{importHelpers}
\alias{importHelpers}
\alias{importClassic2}
\alias{importStages2}
\alias{importClassic21}
\alias{importStages21}
\title{Helper-functions for noninvariance effect size functions}
\usage{
importClassic2(fjson.df)

importStages2(fjson.df)

importClassic21(fjson.df)

importStages21(fjson.df)
}
\arguments{
\item{fjson.df}{flattened json file}
}
\description{
Helper-functions for noninvariance effect size functions
}
\concept{helpers}
