% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scrapePerson.R
\name{scrapePerson}
\alias{scrapePerson}
\title{Scrapes fitbit JSON sleep data from a particular participant in data folder}
\usage{
scrapePerson(idTarget, anon = FALSE, export = FALSE)
}
\arguments{
\item{idTarget}{character of participant (subdirectory) name to scrape}

\item{anon}{whether an anonymized df (and .csv, if export == TRUE) is desired}

\item{export}{whether to export .csv of file to "processed data" subdirectory}
}
\value{

}
\description{
Scrapes fitbit JSON sleep data from a particular participant in data folder
}
