% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scrapeSample.R
\name{scrapeSample}
\alias{scrapeSample}
\title{Scrapes fitbit JSON sleep data from all participants (subdirectories) in a data directory}
\usage{
scrapeSample(anon = FALSE, export = FALSE)
}
\arguments{
\item{anon}{whether an anonymized df (and .csv, if export == TRUE) is desired}

\item{export}{whether to export .csv of file to "processed data" subdirectory}
}
\value{

}
\description{
Scrapes fitbit JSON sleep data from all participants (subdirectories) in a data directory
}
