% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scrapeSample.R
\name{scrapeSample}
\alias{scrapeSample}
\title{Scrapes fitbit JSON sleep data from all participants (subdirectories) in a data directory}
\usage{
scrapeSample(sampbirthdf = NULL, anon = FALSE, export = FALSE, ...)
}
\arguments{
\item{sampbirthdf}{optional character file name (w/ extension) of supplemental data file of id and child birth date (babybirthdarte format: mm/dd/yyyy) in "data" subdirectory}

\item{anon}{whether an anonymized df (and .csv, if export == TRUE) is desired}

\item{export}{whether to export .csv of file to "processed data" subdirectory}

\item{...}{Additional arguments passed to each loop of scrapePerson (e.g., type, sleepdate)}
}
\value{
df of scraped .json sleep data from all subdirectories in data
}
\description{
Scrapes fitbit JSON sleep data from all participants (subdirectories) in a data directory
}
\examples{
\dontrun{
samp.out <- scrapeSample(rawdata = TRUE)
}
}
